// src/modules/teacher-payouts/enums/teacher-payouts.codes.ts
/**
 * Teacher Payout-related error codes (TP_xxx)
 * Range: TP_001 - TP_049
 */
export enum TeacherPayoutErrorCode {
  PAYOUT_NOT_FOUND = 'TP_001',
  PAYOUT_ALREADY_PAID = 'TP_002',
  PAYOUT_INVALID_STATUS_TRANSITION = 'TP_003',
  PAYOUT_ALREADY_EXISTS = 'TP_004',
  PAYOUT_INVALID_TYPE = 'TP_005',
  PAYOUT_AMOUNT_EXCEEDS_REMAINING = 'TP_006',
  PAYOUT_INVALID_AMOUNT = 'TP_007',
  PAYOUT_NOT_FOUND_FOR_CLASS = 'TP_008',
}